Made lots of changes including adding game scaling depending on monitor resolution, as well as a command line flag.

+++ src/dust.c
@@ -82,7 +82,7 @@ void dust_grid_draw(void) {
     
     for (int i = 0; i < data->count; i++) {
         SDL_Color c = pixel_from_index(data->types[i], (int)data->xs[i] + (int)data->ys[i]*gs->gw);
-        const float coeff = 0.5;
+        const f64 coeff = 0.5;
         SDL_SetRenderDrawColor(gs->renderer, c.r*coeff, c.g*coeff, c.b*coeff, c.a);
         SDL_RenderDrawPoint(gs->renderer, data->xs[i], data->ys[i]);
     }
